server/requirements.txt
# LiveKit + OpenAI realtime bridge (tested with these or newer)
livekit-rtc>=0.12.0
livekit-agents>=1.2.0
livekit-plugins-openai>=0.5.0
flask>=3.0.0
python-dotenv>=1.0.1
pytz>=2024.1
